Refactoring out caching from get_next_ip to helper
